refactor: remove opponent
merge opponent into player
