+++ src/main/java/baseball/Player.java
@@ -1,11 +1,14 @@
 package baseball;
 
-// TODO: Add Interface
 public class Player {
 
 	private int[] numbers;
 
 	public Player() {
+	}
+
+	public Player(int[] numbers) {
+		this.numbers = numbers;
 	}
 
 	public void setNumbers(int[] numbers) {
